+++ R/merMod_p.R
@@ -10,7 +10,7 @@
 #'         conditional F-tests with Kenward-Roger approximation for the df (see
 #'         'Details').
 #'
-#' @return A \code{\link{tibble}} with the model coefficients' names (\code{term}),
+#' @return A \code{data.frame} with the model coefficients' names (\code{term}),
 #'         p-values (\code{p.value}) and standard errors (\code{std.error}).
 #'
 #' @details For linear mixed models (\code{lmerMod}-objects), the computation of
@@ -23,6 +23,10 @@
 #'         \cr \cr
 #'         If p-values already have been computed (e.g. for \code{merModLmerTest}-objects
 #'         from the \CRANpkg{lmerTest}-package), these will be returned.
+#'         \cr \cr
+#'         The \code{print()}-method has a \code{summary}-argument, that - in
+#'         case \code{p.kr = TRUE} - also prints information on the approximated
+#'         degrees of freedom (see 'Examples').
 #'
 #' @examples
 #' data(efc)
@@ -38,11 +42,17 @@
 #'
 #' # lme4-fit
 #' library(lme4)
-#' fit <- lmer(Reaction ~ Days + (Days | Subject), data = sleepstudy)
-#' p_value(fit, p.kr = TRUE)
+#' sleepstudy$mygrp <- sample(1:45, size = 180, replace = TRUE)
+#' fit <- lmer(Reaction ~ Days + (1 | mygrp) + (1 | Subject), sleepstudy)
+#' pv <- p_value(fit, p.kr = TRUE)
+#'
+#' # normal output
+#' pv
+#'
+#' # add information on df and t-statistic
+#' print(pv, summary = TRUE)
 #'
 #' @importFrom stats coef pnorm
-#' @importFrom tibble tibble
 #' @importFrom broom tidy
 #' @importFrom dplyr select
 #' @export
@@ -61,7 +71,11 @@ p_value <- function(fit, p.kr = FALSE) {
     se <- unlist(Zelig::get_se(fit))
   } else if (is_merMod(fit)) {
     p <- merMod_p(fit, p.kr)
-    se <- stats::coef(summary(fit))[, 2]
+    sekr <- attr(p, "se.kr", exact = TRUE)
+    if (!is.null(sekr))
+      se <- sekr
+    else
+      se <- stats::coef(summary(fit))[, 2]
   } else if (inherits(fit, c("pglm", "maxLik"))) {
     p <- summary(fit)$estimate[, 4]
     se <- summary(fit)$estimate[, 2]
@@ -85,9 +99,19 @@ p_value <- function(fit, p.kr = FALSE) {
     se <- stats::coef(summary(fit))[, 2]
   }
 
-  tibble::tibble(term = names(p),
-                 p.value = as.vector(p),
-                 std.error = as.vector(se))
+  res <- data.frame(
+    term = names(p),
+    p.value = as.vector(p),
+    std.error = as.vector(se)
+  )
+
+  class(res) <- c("sj_pval", class(res))
+
+  attr(res, "df.kr") <- attr(p, "df.kr", exact = TRUE)
+  attr(res, "se.kr") <- attr(p, "se.kr", exact = TRUE)
+  attr(res, "t.kr") <- attr(p, "t.kr", exact = TRUE)
+
+  res
 }
 
 
@@ -115,11 +139,16 @@ merMod_p <- function(fit, p.kr) {
     #first coefficients need to be data frame
     cs <- as.data.frame(cs)
     # get KR DF
-    df.kr <- suppressMessages(pbkrtest::get_Lb_ddf(fit, lme4::fixef(fit)))
+    df.kr <- get_kr_df(fit)
+    se.kr <- get_kr_se(fit)
+    t.kr <- lme4::fixef(fit) / se.kr
     # compute p-values, assuming an approximate t-dist
-    pv <- 2 * stats::pt(abs(cs$`t value`), df = df.kr, lower.tail = FALSE)
+    pv <- 2 * stats::pt(abs(t.kr), df = df.kr, lower.tail = FALSE)
     # name vector
     names(pv) <- coef_names
+    attr(pv, "df.kr") <- df.kr
+    attr(pv, "se.kr") <- se.kr
+    attr(pv, "t.kr") <- t.kr
   } else {
     message("Computing p-values via Wald-statistics approximation (treating t as Wald z).")
     pv <- 2 * stats::pnorm(abs(cs[, 3]), lower.tail = FALSE)
@@ -132,3 +161,38 @@ merMod_p <- function(fit, p.kr) {
 
   pv
 }
+
+
+#' @importFrom lme4 fixef
+get_kr_df <- function(x) {
+  if (!requireNamespace("pbkrtest", quietly = TRUE))
+    stop("Package `pbkrtest` required for Kenward-Rogers approximation.", call. = FALSE)
+
+  L <- diag(rep(1, length(lme4::fixef(x))))
+  L <- as.data.frame(L)
+  out <- suppressMessages(purrr::map_dbl(L, pbkrtest::get_Lb_ddf, object = x))
+  names(out) <- names(lme4::fixef(x))
+  out
+}
+
+
+#' @importFrom lme4 fixef
+#' @importFrom purrr map_dbl
+get_kr_se <- function(x) {
+  if (!requireNamespace("pbkrtest", quietly = TRUE))
+    stop("Package `pbkrtest` required for Kenward-Roger approximation.", call. = FALSE)
+
+  vcov_adj <- pbkrtest::vcovAdj(x)
+
+  fe <- lme4::fixef(x)
+  le <- length(fe)
+  Lmat <- diag(le)
+
+  se <- purrr::map_dbl(1:le, ~ sqrt(qform(Lmat[.x, ], as.matrix(vcov_adj))))
+  names(se) <- names(fe)
+
+  se
+}
+
+
+qform <- function(x, A) sum(x * (A %*% x))