+++ NEWS.md
@@ -1,3 +1,31 @@
+# sjstats 0.15.0
+
+## General
+
+* Added new Vignette _Statistics for Bayesian Models_.
+
+## New functions
+
+* `equi_test()` to test if parameter values in Bayesian estimation should be accepted or rejected.
+* `mediation()` to print a summary of a mediation analysis from multivariate response models fitted with _brms_.
+
+## Changes to functions
+
+* `link_inverse()` now also returns the link-inverse function for cumulative-family _brms_-models.
+* `model_family()` now also returns an `is_ordinal`-element with information if the model is ordinal resp. a cumulative link model.
+* Functions that access model information (like `model_family()`) now better support `vglm`-models (package _VGAM_).
+* `r2()` now also calculates the standard error for _brms_ or _stanreg_ models.
+* `r2()` gets a `loo`-argument to calculate LOO-adjusted rsquared values for _brms_ or _stanreg_ models. This measure comes conceptionally closer to an adjusted r-squared measure.
+* Effect sizes (`anova_stats()`, `eta_sq()` etc.) are now also computed for mixed models.
+* To avoid confusion, `n_eff()` now computes the number of effective samples, and no longer its ratio in relation to the total number of samples.
+* The column name for the ratio of the number of effective samples in `tidy_stan()` is now named *neff_ratio*, to avoid confusion.
+
+## Bug fixes
+
+* Fixed issue in `se()` for `icc()`-objects, where random effect term could not be found.
+* Fixed issue in `se()` for `merMod`-objects.
+* Fixed issue in `p_value()` for mixed models with KR-approximation, which is now more accurate.
+
 # sjstats 0.14.3
 
 ## General